Layer Details Tabs (#380)
* Add some colors

* comment

* tab
